requirements.txt
GitPython==3.1.43
PyYAML==6.0.1
requests==2.32.3
